Made Events.children optionally callable
There are several cases where Events.children need to be populated on
the go. These changes allow it to be a callable instead of a list. Care
must naturally be taken to ensure that they are not mixed, but the
`children` attribute should only be modified in the instance that
created it.