Change behaviour of auto slider bounds
Previously it was using the distance between value and known bound or
value and zero to set the bounds. However this was problematic when
this distance was zero. Now it simply adds some padding (10). No
guess will be good in general, but a least this one always creates
slideable sliders.