Update signal1d.rst
A number of links to the detailed command descriptions are broken. I don't know if this is a known issue. For example, in this section (Signal1D tools) only the spikes removal tool link actually works. I think it may be due to the presence of a thing called "Signal1DTools" which no longer exists. Singal1D does exist so I'm trying it here. If I'm right I'll do the others but I don't know how to test links within the documentation before I do a pull request, so I need confirmation.
Thanks