Update tools.rst
Thanks. Is this what you mean? As you can tell I'm doing this by picking things that seem to  work elsewhere, without having the slightest knowledge of the syntax. 
I picked that bit of code form the last line of this file:
"Both are available via the :py:class:`~.axes.:py:class:`~.axes.AxesManager`` of the signal."
For which the link works on the site. My problem is not being able to test the code. Is that what "Sphinx" is about? Do I have to learn another thing? Arggh, I'm too old!!