some typos and a suggestion or two
I've put the note I suggested about  in issue #1628 about binning in the general fitting part. It is not specific to Samfire. I have not changed the Samfire part. Not sure if Tomas would prefer to do it. In the change about bounded fitting, I would have put a link to the table, instead of saying "see above", but not sure how to.